==> 86.TemperatureExperiment/02.Select_signal_according_Signals_Distance.Rmd <==
---
title: "AANanopore"
author: "Chao Tang"
date: "2022/7/26"
output: html_document
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(eval      = T, # 在块中运行代码(default = TRUE)
                      highlight = T, # 高亮显示
                      echo      = F, # 是否在输出中包含源代码
                      tidy      = T, # 是否整理代码
                      error     = T, # 是否在输出中包含错误信息
                      warning   = F, # 是否在输出中包含警告(default = TRUE)
                      message   = F, # 是否在输出中包含参考的信息
                      cache.    = F)
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r warning=FALSE}
library(data.table)
library(parallel)
library(caret)
library(IRanges)
library(shiny)
library(plotly)
library(ggpubr)
library(ggplot2)
library(patchwork)
```

```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}
```

```{r}
meta0 <- fread("./data/MetaInfomation/Temperature_Experiment_Meta.txt")
meta0[, file_id := as.character(file_id)]
meta0 <- meta0[file_id %in% gsub(".MainL0.txt", "", list.files("./analysis/86.TemperatureExperiment/SelectedL0"))]
```

# 21 ℃


```{r}
meta <- meta0[temperature == 21]
meta[, sig_file := paste0("./analysis/86.TemperatureExperiment/SelectedL0/", file_id, ".MainL0.txt")]
```

```{r}
sigs <- do.call(rbind, lapply(meta$sig_file, fread))
```

```{r fig.width=10, fig.height=12}
ggplot(sigs, aes(x = Blockade, y = DwellTime, colour = A)) + 
  geom_point() + 
  scale_y_log10() + 
  facet_wrap(~ A, ncol = 1, scales = "free_y")
```

### Distance of signal

#### Background data

```{r}
bgsignals0 <- do.call(rbind, lapply(list.files("./analysis/81.ABFProcessing/SelectedSignals", "_background.txt", full.names = TRUE), fread))[Blockade < 0.3 & DwellTime > 0.3]
bgsignals0[, ID := paste0("Noise_", ID)]
Big_FeatureMatrixs <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/01.SignalDistance/StandardAA_Big_FeatureMatrixs.Rds")
Big_FeatureMatrixs <- Big_FeatureMatrixs[ID %in% bgsignals0$ID]
Big_FeatureMatrixs[, AA := NULL]
```

```{r}
bgsignals <- sigs[A == meta[sample == "blank", file_id], ]
bgsignals[, ID := paste0("Noise_", ID)]

B_file <- bgsignals[, as.character(unique(A))]
B_file <- unlist(strsplit(B_file, "\\."))[1]
Big_FeatureMatrix <- list.files(path = "./analysis/81.ABFProcessing/FeatureMatrix", B_file, full.names = TRUE)
Big_FeatureMatrix <- fread(Big_FeatureMatrix)
Big_FeatureMatrix[, ID := paste0("Noise_", ID)]
Big_FeatureMatrix <- Big_FeatureMatrix[ID %in% bgsignals$ID]
```

```{r}
Big_FeatureMatrixs <- Big_FeatureMatrixs[!ID %in% Big_FeatureMatrix[, ID]]
```

#### Signal data

```{r}
sgsignals <- sigs[A %in% meta[sample != "blank", file_id], ]
B_file <- sgsignals[, as.character(unique(A))]
B_file <- unique(mapply(function(x) unlist(strsplit(x, "\\."))[1], B_file))
Sig_FeatureMatrix <- lapply(B_file, function(x) {
  # print(x)
  fread(list.files(path = "./analysis/81.ABFProcessing/FeatureMatrix", x, full.names = TRUE))[ID %in% sgsignals$ID]
})
Sig_FeatureMatrix <- do.call(rbind, Sig_FeatureMatrix)
```

```{r eval=FALSE}
mclapply(sgsignals[, as.character(unique(A))], function(x) {
  if(file.exists(paste0("./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp21/", x, "_Signals_euclidean_distance.Rds"))) return(NULL)
  SFMi <- Sig_FeatureMatrix[ID %in% sgsignals[A == x, ID]]
  if(nrow(SFMi) > nrow(Big_FeatureMatrix)) {
    FeatureMatrix <- rbind(Big_FeatureMatrix, 
                           Big_FeatureMatrixs[sample(.N, nrow(SFMi) - nrow(Big_FeatureMatrix))], 
                           SFMi)
  } else {
    FeatureMatrix <- rbind(Big_FeatureMatrix[sample(.N, nrow(SFMi))], SFMi)
  }
  setkey(FeatureMatrix, ID)
  FeatureMatrix <- data.frame(FeatureMatrix[, grepl("^X", colnames(FeatureMatrix)) | grepl("DwellTime", colnames(FeatureMatrix)), with = F], row.names = FeatureMatrix[[1]])
  euclidean = stats::dist(FeatureMatrix, method = "euclidean")
  saveRDS(euclidean, file = paste0("./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp21/", x, "_Signals_euclidean_distance.Rds"))
}, mc.cores = 1)
```

## Distance filtering

```{r fig.width=24, fig.height=6}
Run1_dist <- readRDS("./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp21/2023_05_17_0003.1_Signals_euclidean_distance.Rds")
select_sigs_Run1 <- KnnSelecting(signals = sgsignals[A == "2023_05_17_0003.1"], dists = Run1_dist, k = 10, noise = 0)
```

```{r}
subtab <- Sig_FeatureMatrix[ID %in% select_sigs_Run1]
subtab <- data.frame(subtab[, grepl("^X", colnames(subtab)) | grepl("DwellTime", colnames(subtab)), with = F], row.names = subtab[[1]])
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r fig.width=10, fig.height=4}
select_sigs_RemoveOutlier_Run1 <- RemoveOutlier(tree = res.hc, n = 100, Keep = 0.91)
ggplot(sgsignals[ID %in% select_sigs_RemoveOutlier_Run1], aes(x = Blockade, y = DwellTime, colour = A)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  labs(title = length(select_sigs_RemoveOutlier_Run1)) + 
  theme_classic(base_size = 15) + 
  theme(legend.position = "none")
```

```{r}
select_sigs <- sigs[ID %in% c(select_sigs_RemoveOutlier_Run1)]
select_sigs[, A := as.character(A)]
select_sigs <- merge(meta[, .(file_name, amino_acid, temperature, file_id)], select_sigs, by.x = "file_id", by.y = "A")
```

```{r fig.width=10, fig.height=4}
ggplot(select_sigs, aes(x = Blockade, y = DwellTime, colour = file_id)) + 
  geom_point() + 
  scale_y_log10() + 
  facet_wrap(~ file_id, ncol = 1, scales = "free_y")
```

```{r}
fwrite(select_sigs, "./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp21/select_sigs.txt", sep = "\t", quote = FALSE)
```

# 25 ℃


```{r}
meta <- meta0[temperature == 25]
meta[, sig_file := paste0("./analysis/86.TemperatureExperiment/SelectedL0/", file_id, ".MainL0.txt")]
```

```{r}
sigs <- do.call(rbind, lapply(meta$sig_file, fread))
```

```{r fig.width=10, fig.height=12}
ggplot(sigs, aes(x = Blockade, y = DwellTime, colour = A)) + 
  geom_point() + 
  scale_y_log10() + 
  facet_wrap(~ A, ncol = 1, scales = "free_y")
```

### Distance of signal

#### Background data

```{r}
bgsignals0 <- do.call(rbind, lapply(list.files("./analysis/81.ABFProcessing/SelectedSignals", "_background.txt", full.names = TRUE), fread))[Blockade < 0.3 & DwellTime > 0.3]
bgsignals0[, ID := paste0("Noise_", ID)]
Big_FeatureMatrixs <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/01.SignalDistance/StandardAA_Big_FeatureMatrixs.Rds")
Big_FeatureMatrixs <- Big_FeatureMatrixs[ID %in% bgsignals0$ID]
Big_FeatureMatrixs[, AA := NULL]
```

```{r}
bgsignals <- sigs[A == meta[sample == "blank", file_id], ]
bgsignals[, ID := paste0("Noise_", ID)]

B_file <- bgsignals[, as.character(unique(A))]
B_file <- unlist(strsplit(B_file, "\\."))[1]
Big_FeatureMatrix <- list.files(path = "./analysis/81.ABFProcessing/FeatureMatrix", B_file, full.names = TRUE)
Big_FeatureMatrix <- fread(Big_FeatureMatrix)
Big_FeatureMatrix[, ID := paste0("Noise_", ID)]
Big_FeatureMatrix <- Big_FeatureMatrix[ID %in% bgsignals$ID]
```

```{r}
Big_FeatureMatrixs <- Big_FeatureMatrixs[!ID %in% Big_FeatureMatrix[, ID]]
```

#### Signal data

```{r}
sgsignals <- sigs[A %in% meta[sample != "blank", file_id], ]
B_file <- sgsignals[, as.character(unique(A))]
B_file <- unique(mapply(function(x) unlist(strsplit(x, "\\."))[1], B_file))
Sig_FeatureMatrix <- lapply(B_file, function(x) {
  # print(x)
  fread(list.files(path = "./analysis/81.ABFProcessing/FeatureMatrix", x, full.names = TRUE))[ID %in% sgsignals$ID]
})
Sig_FeatureMatrix <- do.call(rbind, Sig_FeatureMatrix)
```

```{r eval=FALSE}
mclapply(sgsignals[, as.character(unique(A))], function(x) {
  if(file.exists(paste0("./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp21/", x, "_Signals_euclidean_distance.Rds"))) return(NULL)
  SFMi <- Sig_FeatureMatrix[ID %in% sgsignals[A == x, ID]]
  if(nrow(SFMi) > nrow(Big_FeatureMatrix)) {
    FeatureMatrix <- rbind(Big_FeatureMatrix, 
                           Big_FeatureMatrixs[sample(.N, nrow(SFMi) - nrow(Big_FeatureMatrix))], 
                           SFMi)
  } else {
    FeatureMatrix <- rbind(Big_FeatureMatrix[sample(.N, nrow(SFMi))], SFMi)
  }
  setkey(FeatureMatrix, ID)
  FeatureMatrix <- data.frame(FeatureMatrix[, grepl("^X", colnames(FeatureMatrix)) | grepl("DwellTime", colnames(FeatureMatrix)), with = F], row.names = FeatureMatrix[[1]])
  euclidean = stats::dist(FeatureMatrix, method = "euclidean")
  saveRDS(euclidean, file = paste0("./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp21/", x, "_Signals_euclidean_distance.Rds"))
}, mc.cores = 2)
```

## Distance filtering

```{r fig.width=24, fig.height=6}
Run1_dist <- readRDS("./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp25/2023_05_16_0006.1_Signals_euclidean_distance.Rds")
select_sigs_Run1 <- KnnSelecting(signals = sgsignals[A == "2023_05_16_0006.1"], dists = Run1_dist, k = 10, noise = 0)
```

```{r}
subtab <- Sig_FeatureMatrix[ID %in% select_sigs_Run1]
subtab <- data.frame(subtab[, grepl("^X", colnames(subtab)) | grepl("DwellTime", colnames(subtab)), with = F], row.names = subtab[[1]])
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r fig.width=10, fig.height=4}
select_sigs_RemoveOutlier_Run1 <- RemoveOutlier(tree = res.hc, n = 100, Keep = 0.91)
ggplot(sgsignals[ID %in% select_sigs_RemoveOutlier_Run1], aes(x = Blockade, y = DwellTime, colour = A)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  labs(title = length(select_sigs_RemoveOutlier_Run1)) + 
  theme_classic(base_size = 15) + 
  theme(legend.position = "none")
```


```{r fig.width=24, fig.height=6}
Run2_dist <- readRDS("./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp25/2023_05_17_0001.1_Signals_euclidean_distance.Rds")
select_sigs_Run2 <- KnnSelecting(signals = sgsignals[A == "2023_05_17_0001.1"], dists = Run2_dist, k = 10, noise = 0)
```

```{r}
subtab <- Sig_FeatureMatrix[ID %in% select_sigs_Run2]
subtab <- data.frame(subtab[, grepl("^X", colnames(subtab)) | grepl("DwellTime", colnames(subtab)), with = F], row.names = subtab[[1]])
res.ds <- stats::dist(subtab)
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r fig.width=10, fig.height=4}
select_sigs_RemoveOutlier_Run2 <- RemoveOutlier(tree = res.hc, n = 100, Keep = 0.91)
ggplot(sgsignals[ID %in% select_sigs_RemoveOutlier_Run2], aes(x = Blockade, y = DwellTime, colour = A)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  labs(title = length(select_sigs_RemoveOutlier_Run2)) + 
  theme_classic(base_size = 15) + 
  theme(legend.position = "none")
```


```{r}
select_sigs <- sigs[ID %in% c(select_sigs_RemoveOutlier_Run1, select_sigs_RemoveOutlier_Run2)]
select_sigs[, A := as.character(A)]
select_sigs <- merge(meta[, .(file_name, amino_acid, temperature, file_id)], select_sigs, by.x = "file_id", by.y = "A")
```

```{r fig.width=10, fig.height=4}
ggplot(select_sigs, aes(x = Blockade, y = DwellTime, colour = file_id)) + 
  geom_point() + 
  scale_y_log10() + 
  facet_wrap(~ file_id, ncol = 1, scales = "free_y")
```

```{r}
fwrite(select_sigs, "./analysis/86.TemperatureExperiment/02.SignalsDistance/Temp25/select_sigs.txt", sep = "\t", quote = FALSE)
```


# Interval time

```{r}
files <- list.files("./analysis/86.TemperatureExperiment/02.SignalsDistance", "select_sigs.txt", full.names = T, recursive = T)

lapply(files, function (fi) {
  sig <- fread(fi)
  sig[, file_id := as.character(file_id)]
  sig[, File := as.character(file_name)]
  sig <- split(sig, sig$file_id)
  SigInterTime <- lapply(sig, function(sigi) {
    print(nrow(sigi))
    setkey(sigi, StartTime)
    ol <- as.data.table(findOverlaps(sigi[, IRanges(StartTime * 10000, EndTime * 10000)], sigi[, IRanges(StartTime * 10000, EndTime * 10000)], type = "within"))[queryHits != subjectHits]
    if(nrow(ol) > 0) {
      ol <- unique(apply(ol, 1, sort)[1, ])
      sigi <- sigi[-c(ol)]
      setkey(sigi, StartTime)
    }
    InterTime <- data.table(File = sigi[, unique(File)], 
                            StartTime = sigi[, EndTime][-c(nrow(sigi))], 
                            EndTime = sigi[, StartTime][-c(1)], 
                            SigBefore = sigi[, ID][-c(nrow(sigi))], 
                            SigAfter = sigi[, ID][-c(1)], 
                            SigBeforeBaseMean = sigi[, BaseMean][-c(nrow(sigi))], 
                            SigAfterBaseMean = sigi[, BaseMean][-c(1)])
    
    filei <- sigi[, unique(File)]
    abf <- readRDS(paste0("./analysis/81.ABFProcessing/ABF/ABF_", filei, ".Rds"))
    abf <- abf[Time >= sigi[, min(StartTime)] & Time <= sigi[, max(EndTime)]]
    L0 <- fread(list.files("./analysis/86.TemperatureExperiment/SelectedL0", sigi[, unique(file_id)], full.names = T))
    ssd <- abf[Sm >= L0[, min(BaseMean)] & Sm <= L0[, max(BaseMean)], sd(pA)]
    abf[, L0 := Sm >= L0[, min(BaseMean) - ssd] & Sm <= L0[, max(BaseMean) + ssd]]
    
    ValidRegion <- IRanges(Rle(abf[, round(mV) == 50 & Sm > 20]))
    Valid <- data.table(satrt = abf[start(ValidRegion), Time], end = abf[end(ValidRegion), Time])
    InterTime <- InterTime[StartTime %inrange% Valid & EndTime %inrange% Valid]
    InterTime2 <- mclapply(seq_len(nrow(InterTime)), function(i) {
      abf[Time > InterTime[i, StartTime] & Time < InterTime[i, EndTime], .(Time1 = diff(range(Time)), Time2 = mean(L0) * diff(range(Time)))]
    }, mc.cores = 10)
    InterTime2 <- do.call(rbind, InterTime2)
    InterTime <- cbind(InterTime, InterTime2)
    return(InterTime)
  })
  SigInterTime <- do.call(rbind, SigInterTime)
  fwrite(SigInterTime, gsub("select_sigs.txt", "select_sigs_IntervalTime.txt", fi), sep = "\t", quote = F, row.names = F)
})
```
